move server files to server directory
