+++ frontend/app/lib/server/validate.server.ts
@@ -4,8 +4,8 @@ import {
   PositionType,
   SlideAnimationType,
   WalletAddressFormatError
-} from './types.js'
-import { isWalletAddress, toWalletAddressUrl } from './utils.js'
+} from '../types.js'
+import { isWalletAddress, toWalletAddressUrl } from '../utils.js'
 
 const rangeError = { message: 'Value has to be between 16 and 24' }
